change all filenames to clearer form (Closes #56). Move error messages to config (Closes #48)
