+++ services/users/src/app.js
@@ -1,8 +1,8 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import cookieParser from 'cookie-parser';
-import userRouter from './routers/userRouter.js';
-import authRouter from './routers/authRouter.js';
+import userRouter from './routers/user.router.js';
+import authRouter from './routers/auth.router.js';
 import { PREFIX } from './config.js';
 
 const app = express();
