+++ services/users/src/routers/auth.router.js
@@ -1,5 +1,5 @@
 import Router from 'express';
-import authController from '../controllers/authController.js';
+import authController from '../controllers/auth.controller.js';
 
 const authRouter = new Router();
