+++ services/users/src/routers/user.router.js
@@ -1,5 +1,5 @@
 import Router from 'express';
-import userController from '../controllers/userController.js';
+import userController from '../controllers/user.controller.js';
 
 const userRouter = new Router();
