+++ services/users/src/services/auth.service.js
@@ -1,9 +1,9 @@
 import bcrypt from 'bcryptjs';
-import tokenService from './tokenService.js';
-import userService from './userService.js';
-import UserDto from '../dtos/user-dto.js';
-import { kafkaTopics } from '../config.js';
-import KafkaNotificationProducerService from './kafkaNotificationProducerService.js';
+import tokenService from './token.service.js';
+import userService from './user.service.js';
+import UserDto from '../dtos/user.dto.js';
+import { kafkaTopics, errorMessages } from '../config.js';
+import KafkaNotifProducer from './kafka-notification-producer.service.js';
 
 class AuthService {
   constructor(notificationProducer) {
@@ -15,13 +15,13 @@ class AuthService {
     try {
       await this.notificationProducer.connect();
     } catch (err) {
-      console.error('Error while connecting to kafka: ' + err);
+      console.error(errorMessages.KAFKA_FAILED_CONNECT + err);
     }
   }
 
   async signUp(userData) {
     const user = await userService.getUserByEmail(userData.email);
-    if (user) throw new Error('User with this email has already exists!');
+    if (user) throw new Error(errorMessages.USER_ALREADY_EXISTS);
 
     const newUser = await userService.createUser(userData);
 
@@ -39,10 +39,10 @@ class AuthService {
 
   async signIn(userData) {
     const user = await userService.getUserByEmail(userData.email);
-    if (!user) throw new Error('User with this email does not exists!');
+    if (!user) throw new Error(errorMessages.USER_NOT_EXISTS_EMAIL);
 
     const isPassEqual = await bcrypt.compare(userData.password, user.password);
-    if (!isPassEqual) throw new Error('Wrong password!');
+    if (!isPassEqual) throw new Error(errorMessages.WRONG_PASSWORD);
 
     const userDto = new UserDto(user);
     const tokens = tokenService.generateTokens(userDto);
@@ -52,11 +52,12 @@ class AuthService {
   }
 
   async refresh(refreshToken) {
-    if (!refreshToken) throw new Error('User is not authorized');
+    if (!refreshToken) throw new Error(errorMessages.NOT_AUTHORIZED);
 
     const userData = tokenService.validateRefreshToken(refreshToken);
     const tokenFromDb = await tokenService.findToken(refreshToken);
-    if (!userData || !tokenFromDb) throw new Error('User is not authorized');
+    if (!userData || !tokenFromDb)
+      throw new Error(errorMessages.NOT_AUTHORIZED);
 
     const user = await userService.getUserById(userData.id);
     const userDto = new UserDto(user);
@@ -67,10 +68,10 @@ class AuthService {
   }
 
   async signOut(refreshToken) {
-    if (!refreshToken) throw new Error('User is not authorized');
+    if (!refreshToken) throw new Error(errorMessages.NOT_AUTHORIZED);
 
     return await tokenService.removeToken(refreshToken);
   }
 }
 
-export default new AuthService(new KafkaNotificationProducerService());
+export default new AuthService(new KafkaNotifProducer());
